Refactor: put _try_to_send_lti_grade in rs_grading.
